More krazy & compile warnings fixes.

+++ libavogadro/src/engines/bsdyengine.cpp
@@ -397,10 +397,10 @@ namespace Avogadro
       connect(m_settingsWidget->showMulti, SIGNAL(stateChanged(int)), this, SLOT(setShowMulti(int)));
       connect(m_settingsWidget->opacitySlider, SIGNAL(valueChanged(int)), this, SLOT(setOpacity(int)));
       connect(m_settingsWidget, SIGNAL(destroyed()), this, SLOT(settingsWidgetDestroyed()));
-      m_settingsWidget->atomRadiusSlider->setValue(10*m_atomRadiusPercentage);
-      m_settingsWidget->bondRadiusSlider->setValue(20*m_bondRadius);
+      m_settingsWidget->atomRadiusSlider->setValue(int(10*m_atomRadiusPercentage));
+      m_settingsWidget->bondRadiusSlider->setValue(int(20*m_bondRadius));
       m_settingsWidget->showMulti->setCheckState((Qt::CheckState)m_showMulti);
-      m_settingsWidget->opacitySlider->setValue(20*m_alpha);
+      m_settingsWidget->opacitySlider->setValue(int(20*m_alpha));
     }
     return m_settingsWidget;
   }
@@ -429,14 +429,12 @@ namespace Avogadro
     setOpacity(settings.value("opacity", 100).toInt());
 
     if (m_settingsWidget) {
-      m_settingsWidget->atomRadiusSlider->setValue(10*m_atomRadiusPercentage);
-      m_settingsWidget->bondRadiusSlider->setValue(20*m_bondRadius);
+      m_settingsWidget->atomRadiusSlider->setValue(int(10*m_atomRadiusPercentage));
+      m_settingsWidget->bondRadiusSlider->setValue(int(20*m_bondRadius));
       m_settingsWidget->showMulti->setCheckState((Qt::CheckState)m_showMulti);
-      m_settingsWidget->opacitySlider->setValue(20*m_alpha);
+      m_settingsWidget->opacitySlider->setValue(int(20*m_alpha));
     }
   }
-
-//   AVOGADRO_ENGINE_FACTORY(BSDYEngine)
 
 }
